Refactor: Introduce ProjectContent sealed interface for type-specific data
This commit refactors the `Project` domain entity by introducing a `ProjectContent` sealed interface. This change encapsulates type-specific fields (for FACE, BODY, MUSCLE, LANDSCAPE content types) into their own data classes (`FaceProjectContent`, `BodyProjectContent`, `MuscleProjectContent`, `LandscapeProjectContent`).

This improves type safety and removes nullable, context-dependent fields from the main `Project` class.

Key changes:
- **Domain:**
    - Replaced flat nullable properties on `Project` with a `content: ProjectContent` property.
    - Updated `AlignFaceUseCase`, `GetCalibrationUseCase`, and `FakeProjectRepository` to use the new `FaceProjectContent` data structure.
- **Data:**
    - Updated `ProjectMapper` to handle mapping between the database entity and the new structured domain model.
    - `ProjectRepositoryImpl` now initializes new projects with a default `FaceProjectContent`.
- **Gradle & Build:**
    - Enabled the `-Xexpect-actual-classes` compiler option across `composeApp`, `domain`, `data`, and `platform` modules.
    - Added a `bundleId` to the iOS framework configuration in `composeApp/build.gradle.kts`.
- **Android:**
    - Migrated `CameraPreview.android.kt` from `LocalLifecycleOwner` (Compose) to `LocalLifecycleOwner` (AndroidX Lifecycle Compose).
    - Updated `NotificationSchedulerImpl.kt` to use the `edit { ... }` ktx extension for SharedPreferences and added missing `@RequiresPermission` annotations.
- **iOS:**
    - Corrected `NSDateComponents` mapping from `kotlinx-datetime` in `NotificationSchedulerImpl.kt`.
    - Adopted `PrimaryTabRow` instead of the deprecated `TabRow` in `StatisticsScreen.kt`.
    - Added `@BetaInteropApi` opt-in to address new warnings in `IosImageCodec.kt` and `IosHomographyTransformer.kt`.

+++ composeApp/src/iosMain/kotlin/com/po4yka/framelapse/platform/BodyPoseDetectorImpl.kt
@@ -172,11 +172,11 @@ class BodyPoseDetectorImpl : BodyPoseDetector {
                             y = 1f - point.location.useContents { y.toFloat() },
                             z = 0f, // Vision doesn't provide Z coordinate
                         ),
-                        confidence = point.confidence.toFloat(),
+                        confidence = point.confidence,
                         isVisible = point.confidence > VISIBILITY_THRESHOLD,
                     ),
                 )
-                totalConfidence += point.confidence.toFloat()
+                totalConfidence += point.confidence
                 keypointCount++
             }
         }
